backend/requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
google-generativeai
chromadb
python-dotenv
PyPDF2
tqdm
